+++ backend/api/projects/constants.py
@@ -1,0 +1,3 @@
+CONTENT_BLOCK_INDEX_WIDTH = 3
+STRING_LIST_MAX_ITEMS = 20
+STRING_LIST_ITEM_MAX_LENGTH = 200
