+++ backend/projects/constants.py
@@ -1,5 +1,3 @@
-"""Константы приложения проектов."""
-
 REFERENCE_SLUG_MAX_LENGTH = 64
 PROJECT_SLUG_MAX_LENGTH = 150
 TITLE_MAX_LENGTH = 150
@@ -15,7 +13,5 @@ ORDER_STEP = 1
 BLOCK_VARIANT_IMAGE_WITH_LIST = 1
 BLOCK_VARIANT_TWO_IMAGES = 2
 BLOCK_VARIANT_IMAGE_WITH_BUTTONS = 3
-CONTENT_BLOCK_INDEX_WIDTH = 3
 
-STRING_LIST_MAX_ITEMS = 20
-STRING_LIST_ITEM_MAX_LENGTH = 200
+
